Move visitors only used by bootstrap
